+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # ImageOptimize Changelog
 
+## 1.6.33 - 2021.04.23
+### Added
+* Added a setting to control the amount an image needs to be scaled down for automatic sharpening to be applied (https://github.com/nystudio107/craft-imageoptimize/issues/263)
+* Added an option to cap the placeholder silhouette SVG size to 32Kb (https://github.com/nystudio107/craft-imageoptimize/issues/270)
+
 ## 1.6.32 - 2021.04.15
 ### Fixed
 * If there are too many sub-folders in an Asset volume, don't display them, return an empty array (https://github.com/nystudio107/craft-imageoptimize/issues/203)